perf: process test results in background with messenger Previously, test results where processed directly in the webhook controller, which kept the request active. The hope is that doing the bulk of the processing in the background instead will lighten up the load on PHP-FPM (and move it to the Messenger processes instead, where it can more easily be distributed without preventing users from accessing the app).
