+++ src/Message/TestingResultRequest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Message;
+
+class TestingResultRequest
+{
+	/**
+	 * @param array<string,mixed> $payload
+	 */
+	public function __construct(
+		private array $payload,
+	) {
+	}
+
+	/**
+	 * @return array<string,mixed>
+	 */
+	public function getPayload(): array
+	{
+		return $this->payload;
+	}
+}
